radeon: one more 32 -> 64 just to be safe
missed this in cadae2a8d9979b2d61eae7d1f7a9ab8f5eb28e83

+++ src/radeon_driver.c
@@ -1309,7 +1309,7 @@ static void RADEONInitMemoryMap(ScrnInfoPtr pScrn)
 	else
 #endif
 	{
-	    uint32_t aper0_base;
+	    uint64_t aper0_base;
 
 	    if (info->ChipFamily >= CHIP_FAMILY_R600) {
 		aper0_base = INREG(R600_CONFIG_F0_BASE);
